Add files via upload
These are samples of Material Quantity take offs done in previous works. Estimates are done through AutoCAD, Bluebeam and Excel. Conduits and other materials are drawn in AutoCAD and summed up using lisp files. Bluebeam is used as a search function for fixtures and Excel is used to give structure to the data.